imprv: make the operator idemotent.
