+++ config/crd/bases/k6.io_k6s.yaml
@@ -2103,6 +2103,7 @@ spec:
                   our runners are in
                 enum:
                 - initialization
+                - wait initialization
                 - initialized
                 - created
                 - started
